move patch query to gql folder
